feat: Add signInWithPhone method to MoonSDK
The code changes add a new method `signInWithPhone` to the `MoonSDK` class. This method allows users to sign in with their phone number and password. It makes use of the `signInWithPassword` function from the `getMoonAuth().auth` object. If there is an error during the sign-in process, it will be thrown.

Note: The suggested commit message adheres to the established conventions of the repository.