feat: Add FormData support to HttpClient
The code changes in `http-client.ts` modify the `HttpClient` class to add support for sending form data in the `FormData` format. Previously, the `FormData` input was not properly handled, resulting in incorrect form data being sent in the request.

The changes include:
- Refactoring the `FormData` handling in the `ContentType.FormData` case to correctly iterate over the keys and values of the `FormData` object.
- Using the `FormData.get()` method to retrieve the property value for each key in the `FormData` object.

These modifications ensure that the form data is correctly appended to the request payload when using the `ContentType.FormData` content type.

Note: The suggested commit message follows the established conventions of the repository.